Added support for other resolutions to conversion.
Fixed issue with gzip files, where buffer for gzip wasn't big enough... I preset the buffer to be the correct size, now, and wrote a function that forces a load of a byte array, using loop, recur and cond. All the goodies!

I added a new namespace, trends.clj, though I don't know if that'll be ready for a bit.