Total rework of the date conversions, to make explicit some of the patterns. The end result is that we now have a way to bucket lat, lon points into MODIS pixels!
Note for the future that in going from modis to lat long, we calculate
the centroid by adding half of the edge-length. There's no need to
subtract this on the way back out, as the call to `int` takes care of
flooring the value back down. Subtracting half of the edge length
would push everything off by half a pixel, always.