No longer forcing reduce by throwing in "?count".
Since we're maxing out mappers on the cluster but also running reducers, we run out of memory. The reduce angle was only needed to keep the output files large, but consolidating tiny ones later wouldn't actually take all that long on a large cluster, so it's not worth the instability on a map slot-heavy cluster.